更新最新的开发规则对项目进行了部分调整，删除了很多东西，因为这些很多是纯粹的套壳，增加了维护工作量收益却很少，弃用后可以解耦代码，并能增加对标准库的认知度，以便以后从事更多的c++开发。

+++ BDXKEngine/Source/Mesh.h
@@ -16,9 +16,9 @@ public:
 	}
 	int GetTrianglesCount()
 	{
-		return triangles.size();
+		return (int)triangles.size();
 	}
-	std::vector <int> GetTriangles()
+	std::vector <UINT32> GetTriangles()
 	{
 		return { triangles };
 	}
@@ -50,7 +50,7 @@ public:
 		}
 	}
 
-	void SetTriangles(std::vector <int> data)
+	void SetTriangles(std::vector <UINT32> data)
 	{
 		size_t size = data.size();
 		if (size % 3 != 0)
@@ -97,7 +97,7 @@ public:
 	}
 private:
 	std::vector<Vertex> vertices;
-	std::vector<int> triangles;
+	std::vector<UINT32> triangles;
 
 	CComPtr<ID3D11Buffer> vertexBuffer;
 	CComPtr<ID3D11Buffer> triangleBuffer;
